+++ package.json
@@ -5,6 +5,7 @@
   "main": "index.js",
   "scripts": {
     "start": "webpack-dev-server --open --progress",
+    "start:test": "webpack-dev-server --config webpack.test.config.js --open --progress",
     "build": "webpack --config webpack.config.js --progress",
     "watch": "webpack --watch",
     "test": "echo \"Error: no test specified\" && exit 1",
