+++ src/components/userContent/cardTemplate/orderItem/orderItem.js
@@ -4,8 +4,10 @@ import orderItem from './orderItem.hbs';
 export default class OrderItem {
   render(target, props) {
     let orderItemTemplate = orderItem(props);
+
     let fieldForOrderItem = document.createElement('div');
     fieldForOrderItem.innerHTML = orderItemTemplate;
+
     target.appendChild(fieldForOrderItem);
   }
 }
