+++ src/index.html
@@ -7,10 +7,8 @@
 </head>
 <body>
 <main id="main-screen">
-    <h1>Hello</h1>
-    <script id="screen" type="text/x-handlebars-template">
-        {{screen}}
-    </script>
+    <div id="card-test">
+    </div>
 </main>
 </body>
 </html>
