+++ src/test.html
@@ -1,10 +1,14 @@
 <!DOCTYPE html>
-<html lang="en">
+<html>
 <head>
-    <meta charset="UTF-8">
-    <title>$Title$</title>
+    <meta charset="utf-8">
+    <title>Exadel Food Delivery</title>
+    <link href="https://fonts.googleapis.com/css?family=Tajawal" rel="stylesheet">
 </head>
 <body>
-$END$
+<main id="main-screen">
+    <div id="card-test">
+    </div>
+</main>
 </body>
 </html>
